tail_b.rs hands the computed-call arm and the scalar tail to tail_c.rs, a pure text move under the 800-line ceiling
