+++ crates/almide-mir/src/lower/tail.rs
@@ -803,3 +803,4 @@ impl LowerCtx {
 }
 
 include!("tail_b.rs");
+include!("tail_c.rs");
